Finished the PaymentMethod component
Decided to only offer credit card payment, as that is what was required
by the client (eg. Proffessor).

The `setJson` argument should be a function  that saves the content of
the card information. The content is returned in Json format so it can
be saved in the backend and verifyied later.